Add security to management of processes through supervisord
As part of this commit we do following:
a) Allow supervisorctl to connect to supervisord only through unix
domain sockets. We also make the permissions of supervisord unix
domain sockets restrictive so that only root can invoke supervisorctl
command.
b) Make changes in nodemgr so that we do not depend on Juniper
proprietary extensions to supervisord.

Change-Id: Ie0315a12b0a425d026a386c1441c666e2cc43223
Closes-PR: 1383939
Closes-PR: 1383503