[Server manager : Interface setup for multi-interface is broken Edit]
Closes-Bug : #1402885

Issue:
For setting up normal and bond interface we use to call this script.

/opt/contrail/bin/setup-vnc-interfaces --device eth1 --ip 192.168.100.4/24

I think post 1.2 Ignatious had done re-organized the code to have a
wrapper script which in turn calls the corresponding script.
The wrapper scripts seem to have dependency on fab.
Which will not be installed on target and because of that interface is
not setup at all.

Proposed Fix:
Call the original script instead, which is at

/usr/local/lib/python2.7/dist-packages/contrail_provisioning/common/interface_setup.py

Testing:
Issued provision with multiple interface, saw the interface getting
created and provisoning complete.

Change-Id: I5d767f0cd3f70e5adab9be63945a33999200351b