SM-Storage: Add recovery of ceph-monitor in case ceph-mon doesn't comes up
Issue:
because of some reason ceph-mon was not coming up on couple of nodes. It
stays in the same state even re-provisioning doesn't help.

Root-cause:
Root cause of ceph-mon not coming up is not exactly known. For some
reason "ceph-mon --mkfs" was creating the monitor correctly and causing
monitor to fail to come up. On next run, it just checks if
/var/lib/ceph/mon/<id> directory exists, if yes, it doesn't try to
recover or check if all required files are created or not.
Puppet it should be recover from that state and retry again

Fix:
Added a additional check for {mon_data}/done file, these are created
after "ceph-mon --mkfs" is successful.

Unit Test:
Tried the patch in problematic setup and checked that ceph-mon came by
itself. No manual intervention was required.

Partial-Bug: #1423618

Change-Id: Icf6a5623416c5f8b0b082aa0cbe5b3a12e6cf266