Server Manager: provision complete status is lost when compute role is present Closes-Bug: #1424870
provision status is updated by a curl request to smgr.
On a compute node, Reboot is issued and then it passes to provision
completed.
since reboot is not blocking.

There is a possibility that curl request could be sent or not sent,
based on when reboot completes.

Fix:
Move reboot to provision complete stage.
and in compute just mark for reboot.

Also move the openstack related post operations to a new resource.
Also fix the report_status to run only once , compute_completed status

Change-Id: Ie1b6db744faf3b33dd4711b491fd9352fe4ad31b