Closes-Bug: #1469264 - Neutron notification_driver param set incorrectly.
Previously it was being set to empty if ceilometer was not enabled.
Patch 2:
- Reverting change to set this config for Ceilometer
- Whatever is the default value when neutron is installed, that will be taken
Patch 3:
- Missed removing one line

Change-Id: I557d7e3bcb8a4d9002fe65ccdc015212534055dd

Closes-Bug: #1478101 - Heat Auth Encryption Key set incorrectly

Previously the auth_encryption_key param of heat engine was being set as empty.
This is caused heat stack-create to fail.
Fab provisioned Heat uses Default auth key (This param is not set).

Fix:
Since puppet manifest of heat engine requires auth key as mandatory param, we are setting it to the default that Heat internally uses
If we wish to change the auth key, we can provide it in the openstack params as ::openstack::config::heat_encryption_key
There is an accompanying check-in to server-manager-repo to set this param there

Change-Id: Ie2c828fff0a206eec5e963a6902e9413d56026ac

Related-Bug: #1471596 - Smgr changes for Keepliaved independence for Cluster monitor and monitoring jobs

This check-in:
Is the Smgr side changes made for the checkins at:
https://review.opencontrail.org/#/c/12280/
and https://review.opencontrail.org/#/c/12281/

Extra parameters are being passed to the script setup-vnc-galera
These parameters are set or configured in ha_config.pp

Should go in after Thilak's commit:
https://review.opencontrail.org/#/c/12734/4

Small bug fix - Verified setup-vnc-galera ran on HA Setup

Change-Id: I9263248e2f632fc04dbebbb99afbd0abeff45390

Related-Bug: #1471596 - Readding Fab to Server Manager Provisioning

This check-in:
Brings back Fab package to Server Manager provisioning
There are many contrail-provisioning scripts that use fab commands and sm puppet scripts have to be extricated from these before fab can be removed

Change-Id: Ib289948e0e0317cff1a9de064e86b22b6acca265