Closes-Bug: #1478101 - Heat Auth Encryption Key set incorrectly
Previously the auth_encryption_key param of heat engine was being set as empty.
This is caused heat stack-create to fail.
Fab provisioned Heat uses Default auth key (This param is not set).

Fix:
Since puppet manifest of heat engine requires auth key as mandatory param, we are setting it to the default that Heat internally uses
If we wish to change the auth key, we can provide it in the openstack params as ::openstack::config::heat_encryption_key
There is an accompanying check-in to server-manager-repo to set this param there

Change-Id: Ie2c828fff0a206eec5e963a6902e9413d56026ac