Closes-Bug: #1494525 SM: Upgrade from R2.20 Build #64 to R2.21 build #96 fails due to Keepalived package dependency
Issue:
keepalived in 2.21 needs to be an older version.
so puppet latest doesnt upgrade it to the latest.
contrail-openstack-config has a dependancy on that
and hence it upgrade of these packages fails.

Fix:
Install the specific version of keepalived

Testing:
After upgrade make sure that the package are the latest in
contrail-versiond and dpkg -l | grep contrail

Change-Id: I12d4b3f094e95da228fd04e0df6f1491ff04c783