Closes-Bug: #1532077 Server Manager: compute node provisioning fails at kernel upgrade.
Issue:
Grub will have default kernel to be started as the latest kernel.
In 14.04.2,the kernel version is 3.16.0-30 which is latest than
3.13.0-40.

In this case an explicity entry has to be made at gurb.cfg to set
the kernel to boot with.

This change was done in Fab recently.
https://github.com/Juniper/contrail-fabric-utils/commit/df8ba5ca6b3226b8841612c29a17463e9e8a0580

Patch set 2)
1)update-grub is ideompotent , so need to make an entry in exec.out and
check for it.

Testing:
Tested on the setup where the issue was seen and see provision
completed.

Change-Id: I304b94694ad607a72f98746be04e660a8bb51076