Closes-Bug: #1524192
Issue:
1)when contrail-openstack-config is installed rabbimt-server
starts with init.d and not through supervisor
The restart of supervisor was not capable of restarting the rabbitmq
as that instance of rabbitmq was not managed by supervisor.

This issue was not seen always as ,If supervisor files of
support-services could be put before rabbimq-server comes up for the
first time.
2)rabbit-env.conf was not populated properly.

3)
provision_encap_priority was broken and was failing,
It was not being caught by provision status as
config_started and config_completed was not covering this.

Fix:
Explicitly stop rabbitmq, kill epmd and restart supervisor-support
service.

Testing:
Provisioned on the node were the issue was seen
and made sure rabbitmq was up, and rabbitmqctl cluster_status
show all the nodes.

Change-Id: If9fc563364101e1d05740c667099debebd09003e