Closes-Bug: #1488259 Issue: Puppetlabs reboot seems to have inconsistent behavior. From old debugs we were able to see refresh events were not recieved by the reboot resource.
Changes Done:
1) Add more debugs to reboot type.
2) Bring latest reboot module.
3) If reboot  right after kernel upgrade fails,
Check if kernel is upgraded at compute and if not reboot the target,
so that while installing vrouter, correct vrouter gets installed.

Testing:
1)Check in a single-node setup that reboot-happens after kernel-upgrade.

Conflicts:

	contrail/environment/modules/contrail/manifests/compute.pp
	contrail/environment/modules/contrail/manifests/compute/install.pp

Change-Id: I17f3bd2872d10fbf8adf134696623babdd4481cd

Closes-Bug: #1532077
Server Manager: compute node provisioning fails at kernel upgrade.

Issue:
Grub will have default kernel to be started as the latest kernel.
In 14.04.2,the kernel version is 3.16.0-30 which is latest than
3.13.0-40.

In this case an explicity entry has to be made at gurb.cfg to set
the kernel to boot with.

This change was done in Fab recently.
https://github.com/Juniper/contrail-fabric-utils/commit/df8ba5ca6b3226b8841612c29a17463e9e8a0580

Patch set 2)
1)update-grub is ideompotent , so need to make an entry in exec.out and
check for it.

Testing:
Tested on the setup where the issue was seen and see provision
completed.

Change-Id: I304b94694ad607a72f98746be04e660a8bb51076