Partial-Bug: #1543333 - Fix for Upgrade of 3.0 2709 kilo -> 3.0 2711 kilo
Failure was in config of 2711 kilo.
The failure was because the rabbitmq server was not getting restarted properly.
Previously supervisor-support-service service was not being controlled by puppet.
It was in script form_rqm_cluster.
Now have introduced that service as puppet resource, it is being notified when verify_rabbitmq runs.

Rabbit-server restarts correctly during upgrade and provision completed.

Change-Id: I3a844694f71d5f139dd3cd8bc4c3665b29f5c763