Closes-Bug: #1620866 - Option to Separate Config and Analytics Dbs
If config node doesn't have db role, it is provisioned with cassandra and zookeeper to form an independent cluster on all config nodes
To do this, we need to add the following option in the cluster JSON:
            "parameters" : {
                "provision": {
                    "contrail": {
                        "config": {
                            "config_manage_db": true
                        }
                    }
                }
            }

Also note:

Database role still HAS to be given on the collector nodes or on any other non-config nodes in the cluster. The analytics nodes will use these nodes as its database nodes.
If database role is configured on the config nodes, then both config and analytics will use the same set of database nodes.

Change-Id: I9137d9bcd082efdf09de8da56bb2ccd1ba39c6d9